+++ CHANGELOG.md
@@ -5,11 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [Unreleased] - 2022-02-20
 
 ### Added
 
 - Support 0-9 for roadside stickers
+- Support seasonal & special decor: Nintendo Account, Roadside Winter Stickers, Lunar New Year
+- 3 views which show various levels of detail for decor collections: Simple, Advanced, Exhaustive
+- Support multiple colors for roadside stickers (exhaustive view)
 - Proper versioning and error handling for exporting/import collections
 
 ### Changed
@@ -50,7 +53,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Sharing via URL
 - "About" section
 
-[unreleased]: https://github.com/Notuom/pikmin-bloom-checklist/compare/v1.1.0...HEAD
+[unreleased]: https://github.com/Notuom/pikmin-bloom-checklist/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/Notuom/pikmin-bloom-checklist/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/Notuom/pikmin-bloom-checklist/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/Notuom/pikmin-bloom-checklist/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/Notuom/pikmin-bloom-checklist/releases/tag/v1.0.0